Small updates to improve numerical stability and identify operations that could cause falures
