+++ examples/mountain_car.py
@@ -7,29 +7,29 @@ import tensorflow as tf
 #from tensorflow import logging
 np.random.seed(0)
 
-from pilco.utils import policy, rollout
+from pilco.utils import policy, rollout, Normalised_Env
 
-class Normalised_Env():
-    def __init__(self, env_id, m, std):
-        self.env = gym.make(env_id).env
-        self.action_space = self.env.action_space
-        self.observation_space = self.env.observation_space
-        self.m = m
-        self.std = std
-
-    def state_trans(self, x):
-        return np.divide(x-self.m, self.std)
-
-    def step(self, action):
-        ob, r, done, _ = self.env.step(action)
-        return self.state_trans(ob), r, done, {}
-
-    def reset(self):
-        ob =  self.env.reset()
-        return self.state_trans(ob)
-
-    def render(self):
-        self.env.render()
+# class Normalised_Env():
+#     def __init__(self, env_id, m, std):
+#         self.env = gym.make(env_id).env
+#         self.action_space = self.env.action_space
+#         self.observation_space = self.env.observation_space
+#         self.m = m
+#         self.std = std
+#
+#     def state_trans(self, x):
+#         return np.divide(x-self.m, self.std)
+#
+#     def step(self, action):
+#         ob, r, done, _ = self.env.step(action)
+#         return self.state_trans(ob), r, done, {}
+#
+#     def reset(self):
+#         ob =  self.env.reset()
+#         return self.state_trans(ob)
+#
+#     def render(self):
+#         self.env.render()
 
 SUBS = 5
 T = 25
@@ -70,6 +70,10 @@ ep_rewards = np.zeros((len(X)//T,1))
 
 for i in range(len(ep_rewards)):
     ep_rewards[i] = sum(all_Rs[i * T: i*T + T])
+
+for model in pilco.mgpr.models:
+    model.likelihood.variance.assign(0.05)
+    set_trainable(model.likelihood.variance, False)
 
 r_new = np.zeros((T, 1))
 for rollouts in range(5):
